+++ src/Interval.java
@@ -1,30 +1,30 @@
 public class Interval {
-    public double min;
-    public double max;
+    public float min;
+    public float max;
 
     public Interval() {
         min = Util.INFINITY;
         max = Util.NEGATIVE_INFINITY;
     }
 
-    public Interval(double min, double max) {
+    public Interval(float min, float max) {
         this.min = min;
         this.max = max;
     }
 
-    double size() {
+    float size() {
         return max - min;
     }
 
-    boolean contains(double x) {
+    boolean contains(float x) {
         return x >= min && x <= max;
     }
 
-    boolean surrounds(double x) {
+    boolean surrounds(float x) {
         return x > min && x < max;
     }
 
-    double clamp(double x) {
+    float clamp(float x) {
         if(x < min) return min;
         if(x > max) return max;
         return x;
